refactor session / formula

+++ Assets/Scripts/GameModel/Entity/AttributeEntity.cs
@@ -3,36 +3,36 @@ using UnityEngine;
 public interface IEffectAttribute
 {
     int PowerAddition { get; }
-    int PowerRatio { get; }
+    float PowerRatio { get; }
     int NormalDamageAddition { get; }
-    int NormalDamageRatio { get; }
+    float NormalDamageRatio { get; }
 
-    void ApplyModify(EffectAttributeType type, int value);
+    void ApplyModify(EffectAttributeType type, float value);
 }
 
 public class CardPlayAttributeEntity : IEffectAttribute
 {
     public int PowerAddition { get; private set; } = 0;
-    public int PowerRatio { get; private set; } = 0;
+    public float PowerRatio { get; private set; } = 0f;
     public int NormalDamageAddition { get; private set; } = 0;
-    public int NormalDamageRatio { get; private set; } = 0;
+    public float NormalDamageRatio { get; private set; } = 0f;
 
     public CardPlayAttributeEntity()
     {
     }
 
-    public void ApplyModify(EffectAttributeType type, int value)
+    public void ApplyModify(EffectAttributeType type, float value)
     {
         switch (type)
         {
             case EffectAttributeType.PowerAddition:
-                PowerAddition += value;
+                PowerAddition += (int)value;
                 break;
             case EffectAttributeType.PowerRatio:
                 PowerRatio += value;
                 break;
             case EffectAttributeType.NormalDamageAddition:
-                NormalDamageAddition += value;
+                NormalDamageAddition += (int)value;
                 break;
             case EffectAttributeType.NormalDamageRatio:
                 NormalDamageRatio += value;
